# apps/interlabs-crm-demo/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Interlab Portal API
  version: 1.0.0
  description: |
    Internal Enterprise Dashboard — CRM + ERP + Realtime Operations Hub for
    PT. Interlab Sentra Solutions Indonesia.

    Covers: Auth, Users & Roles, PO lifecycle (11 stages across 5 divisions),
    Sales, Admin & Log, Finance, Technical, HRGA/Legal, Tax & Insurance,
    Notifications, Chat, Email Templates, Files.

    ## Response envelope
    Every response uses a consistent envelope:
      success: `{ "success": true,  "data": ..., "meta"?: {...} }`
      error:   `{ "success": false, "error": "message", "code"?: "..." }`

    ## RBAC
    Every protected endpoint declares `x-rbac` with the feature_key,
    capability, allowed role_keys, and row-level scope. The RBAC middleware
    enforces `(feature, capability)` against `role_permissions` on every
    request; `scope` is applied to queries via the sameRoleScope /
    row-level-security layer.

    ## PO lifecycle side-effects
    Endpoints marked `x-po-side-effects` MUST, in a single transaction:
      1. Append to `purchase_order_status_history`
      2. Append to `purchase_order_tracking_events`
      3. Update `purchase_orders.current_status`
      4. Emit the declared notification templates (if enabled)
  contact:
    name: Interlab Portal Engineering
    email: zakyindrasatriaputra@gmail.com
servers:
  - url: https://api.interlab-portal.com
    description: Production
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Email + password login (reCAPTCHA v2 required)
      security: []
      description: |
        Rate limited to 5 attempts per IP per 15 minutes and per email.
        reCAPTCHA v2 token is verified server-side before password check.
        Returns access_token (1h) + refresh_token (7d, or 30d with remember_me).
      x-rate-limit: 5-per-15min-per-ip-and-email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/LoginResponseData' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate current refresh token
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange refresh token for new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/RefreshTokenResponseData' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ====================================================================
  # Users  (same-role-scoped for non-Superadmin/CEO)
  # ====================================================================
...
  /api/admin-log/awb:
    get:
      tags: [Admin & Log]
      summary: List AWB records
      x-rbac: { feature: awb, capability: view_own, allowed_roles: [superadmin, ceo, admin_log, technical, finance], scope: own }
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - { in: query, name: current_awb_status, schema: { type: string, enum: [Registered, Processed, Arrived] } }
        - { in: query, name: related_po_id, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Paginated AWBs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AWBRecord' }
    post:
      tags: [Admin & Log]
      summary: Create AWB
      description: |
        If awb_tracking_number is provided at creation, PO status auto-advances
        to Shipped (fires admin_log.awb.shipped).
      x-rbac: { feature: awb, capability: create, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      x-po-side-effects:
        notification_templates: [admin_log.awb.shipped]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AWBCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/AWBRecord' }
  /api/admin-log/awb/{id}:
    parameters: [ { $ref: '#/components/parameters/UUIDPathId' } ]
    get:
      tags: [Admin & Log]
      summary: Get AWB
      x-rbac: { feature: awb, capability: view_own, allowed_roles: [superadmin, ceo, admin_log, technical, finance], scope: own }
      responses:
        '200':
          description: AWB
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/AWBRecord' }
    put:
      tags: [Admin & Log]
      summary: Update AWB
      x-rbac: { feature: awb, capability: edit, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AWBCreateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/AWBRecord' }
    delete:
      tags: [Admin & Log]
      summary: Soft-delete AWB
      x-rbac: { feature: awb, capability: delete, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      responses:
        '204': { description: Deleted }
...
  /api/admin-log/operational:
    get:
      tags: [Admin & Log]
      summary: List operational records
      x-rbac: { feature: admin_operational, capability: view_own, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - { in: query, name: reporting_month, schema: { type: string, format: date, description: 'YYYY-MM-01' } }
        - { in: query, name: expense_status, schema: { type: string, enum: [Pending, Paid, Cancelled] } }
      responses:
        '200':
          description: Paginated records
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AdminOperationalRecord' }
    post:
      tags: [Admin & Log]
      summary: Create operational record
      x-rbac: { feature: admin_operational, capability: create, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminOperationalCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/AdminOperationalRecord' }
  /api/admin-log/operational/{id}:
    parameters: [ { $ref: '#/components/parameters/UUIDPathId' } ]
    get:
      tags: [Admin & Log]
      summary: Get operational record
      x-rbac: { feature: admin_operational, capability: view_own, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      responses:
        '200':
          description: Record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/AdminOperationalRecord' }
    put:
      tags: [Admin & Log]
      summary: Update operational record
      x-rbac: { feature: admin_operational, capability: edit, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminOperationalCreateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/AdminOperationalRecord' }
    delete:
      tags: [Admin & Log]
      summary: Soft-delete operational record
      x-rbac: { feature: admin_operational, capability: delete, allowed_roles: [superadmin, ceo, admin_log], scope: own }
      responses:
        '204': { description: Deleted }

  # ====================================================================
  # Finance — PO Customer (auto-created from Sales PO submit)
  # ====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (1h expiry). Obtained from `POST /api/auth/login`.
        Refresh with `POST /api/auth/refresh` (refresh token 7d, or 30d with remember_me).
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      in: query
      name: limit
      required: false
      schema: { type: integer, enum: [25, 50, 75, 100], default: 25 }
    SearchParam:
      in: query
      name: search
      required: false
      schema: { type: string }
...
    UUIDPathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ValidationError:
      description: Request body / params failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    RateLimited:
      description: Too many requests (login rate limit, etc.)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code: { type: string }
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }

    # ------------------------------------------------------------------
    # Enums
    # ------------------------------------------------------------------
    RoleKey:
      type: string
      enum: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
...
    Currency:
      type: string
      enum: [IDR, USD, EUR]
...
    AuditFields:
      type: object
      properties:
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
        created_by: { type: string, format: uuid, nullable: true, readOnly: true }
        updated_by: { type: string, format: uuid, nullable: true, readOnly: true }
        deleted_at: { type: string, format: date-time, nullable: true, readOnly: true }
...
    LoginRequest:
      type: object
      required: [email, password, recaptcha_token]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        recaptcha_token:
          type: string
          description: reCAPTCHA v2 token, verified server-side
        remember_me:
          type: boolean
          default: false
          description: Extends refresh-token expiry to 30 days
    LoginResponseData:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, description: Seconds until access token expiry }
        user: { $ref: '#/components/schemas/User' }
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    RefreshTokenResponseData:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer }

    # ------------------------------------------------------------------
    # User
    # ------------------------------------------------------------------
    User:
      type: object
      required: [id, email, role, display_name, account_status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/RoleKey' }
        permission_level: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        display_name: { type: string }
        account_status:
          type: string
          enum: [active, inactive, suspended]
        managed_role_scope: { type: string, nullable: true }
        can_manage_same_role: { type: boolean }
        feature_permission_scope: { type: string, nullable: true }
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
...
    AWBRecord:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
        - type: object
          properties:
            id: { type: string, format: uuid, readOnly: true }
            awb_record_number: { type: string, readOnly: true }
            related_po_id: { type: string, format: uuid }
            related_po_number: { type: string, nullable: true, readOnly: true }
            customer_id: { type: string, format: uuid, nullable: true }
            supplier_or_manufacturer: { type: string, nullable: true }
            forwarder_or_courier: { type: string, nullable: true }
            awb_tracking_number: { type: string, nullable: true, description: "Trigger field 1 — entering this sets PO status to Shipped" }
            shipment_method: { type: string, enum: [Air, Sea, Land, Courier], nullable: true }
            origin_country: { type: string, nullable: true }
            transit_country_or_hub: { type: string, nullable: true }
            destination: { type: string, nullable: true }
            despatch_date: { type: string, format: date, nullable: true }
            transit_date: { type: string, format: date, nullable: true, description: "Trigger field 2 — entering this sets PO status to Customs" }
            arrival_date: { type: string, format: date, nullable: true, description: "Trigger field 3 — entering this sets PO status to Arrived" }
            current_awb_status: { type: string, enum: [Registered, Processed, Arrived] }
            weight_kg: { type: number, nullable: true }
            package_count: { type: integer, nullable: true }
            description_of_goods: { type: string, nullable: true }
            incoterm: { type: string, nullable: true }
            notes: { type: string, nullable: true }
            attachment_awb_file_ids:
              type: array
              items: { type: string, format: uuid }
            attachment_invoice_file_ids:
              type: array
              items: { type: string, format: uuid }
            attachment_payment_file_ids:
              type: array
              items: { type: string, format: uuid }
    AWBCreateRequest:
      type: object
      required: [related_po_id]
      properties:
        related_po_id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        supplier_or_manufacturer: { type: string }
        forwarder_or_courier: { type: string }
        awb_tracking_number: { type: string }
        shipment_method: { type: string, enum: [Air, Sea, Land, Courier] }
        origin_country: { type: string }
        transit_country_or_hub: { type: string }
        destination: { type: string }
        despatch_date: { type: string, format: date }
        transit_date: { type: string, format: date }
        arrival_date: { type: string, format: date }
        weight_kg: { type: number }
        package_count: { type: integer }
        description_of_goods: { type: string }
        incoterm: { type: string }
        notes: { type: string }
        attachment_awb_file_ids:
          type: array
          items: { type: string, format: uuid }
        attachment_invoice_file_ids:
          type: array
          items: { type: string, format: uuid }
        attachment_payment_file_ids:
          type: array
          items: { type: string, format: uuid }
...
    AdminOperationalRecord:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
        - type: object
          properties:
            id: { type: string, format: uuid, readOnly: true }
            operational_record_number: { type: string, readOnly: true }
            reporting_month: { type: string, format: date, description: First day of the reporting month }
            department: { type: string, nullable: true }
            expense_category: { type: string, nullable: true }
            expense_subcategory: { type: string, nullable: true }
            transaction_date: { type: string, format: date, nullable: true }
            period_start: { type: string, format: date, nullable: true }
            period_end: { type: string, format: date, nullable: true }
            vendor_or_payee: { type: string, nullable: true }
            related_po_id: { type: string, format: uuid, nullable: true }
            description: { type: string, nullable: true }
            currency: { $ref: '#/components/schemas/Currency' }
            amount: { type: number, nullable: true }
            payment_method: { type: string, enum: [Cash, Transfer, Credit Card], nullable: true }
            expense_status: { type: string, enum: [Pending, Paid, Cancelled] }
            workflow_status: { type: string, enum: [draft, submitted, reviewed] }
            notes: { type: string, nullable: true }
            attachment_report_file_ids:
              type: array
              items: { type: string, format: uuid }
            attachment_supporting_file_ids:
              type: array
              items: { type: string, format: uuid }
    AdminOperationalCreateRequest:
      type: object
      required: [reporting_month]
      properties:
        reporting_month: { type: string, format: date }
        department: { type: string }
        expense_category: { type: string }
        expense_subcategory: { type: string }
        transaction_date: { type: string, format: date }
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        vendor_or_payee: { type: string }
        related_po_id: { type: string, format: uuid }
        description: { type: string }
        currency: { $ref: '#/components/schemas/Currency' }
        amount: { type: number }
        payment_method: { type: string, enum: [Cash, Transfer, Credit Card] }
        expense_status: { type: string, enum: [Pending, Paid, Cancelled] }
        notes: { type: string }
        attachment_report_file_ids:
          type: array
          items: { type: string, format: uuid }
        attachment_supporting_file_ids:
          type: array
          items: { type: string, format: uuid }

    # ------------------------------------------------------------------
    # Finance — PO Customer, PR, Invoice Manufacture, Invoice Customer
    # ------------------------------------------------------------------
...